requirements.txt
twint
nest_asyncio
python-telegram-bot
datetime
aiohttp
aiodns
beautifulsoup4
cchardet
dataclasses
elasticsearch
pysocks
pandas>=0.23.0
aiohttp_socks<=0.4.1
schedule
geopy
fake-useragent
googletransx